backend/requirements.txt
# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
websockets>=11.0.3

# Database
motor==3.2.0
pymongo==4.4.0  # Updated to fix import error

# Authentication & Security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.24.1

# AI Components
opencv-python>=4.8.0
tensorflow>=2.13.0
onnx>=1.14.0
onnxruntime>=1.15.0
librosa>=0.10.0
numpy>=1.24.0
pillow>=10.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
